requirements-cloud.txt
# Streamlit Cloud Optimized Requirements
# Use this for Streamlit Cloud deployment to reduce package size

# Install project as editable package
-e .

# Core AI/ML Libraries (CPU-only versions for faster deployment)
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
transformers>=4.30.0

# NLP & Trend Analysis
pytrends>=4.9.0
trendspy>=0.1.6
trendspyg>=0.2.0
jieba>=0.42.1
scikit-learn>=1.3.0

# LLM Integration
openai>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0

# Web Application
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
